Split expression parser
